requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.11
langchain-core>=0.1.0
chromadb>=0.4.0
python-dotenv>=1.0.0
pandas>=1.5.0
pandera>=0.18.0
pyyaml>=6.0
sqlite3